requirements.txt
# Andrew Teesdale's Added Requirements
asgiref==3.8.1
astroid==3.3.5
beautifulsoup4==4.12.3
dill==0.3.9
Django==4.2.16
django-bootstrap-v5==1.0.11
django-core==1.4.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyenchant==3.2.2
pylint==3.3.1
soupsieve==2.6
sqlparse==0.5.1
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
